Fixed lirc script errors when MCE device is connected
The DEVICES variable must be quoted to preserve newlines in its content.
Without them everything is squashed down to a single line, and every 'eventN'
device is returned in the grep output, instead of just the MCE device.  This
lead to syntax errors in the test expressions on every lircd start/stop.