Make sure all configuration files can be resolved.
Normally autofs mounts shares early in the boot process, whenever a file is requested. However, this automatic behavior doesn't work on FreeBSD NFS. Programs like Lirc will therefor fail when relying on configuration files stored (and symlinked) to a FreeBSD NFS shares.  This loop will check if the configuarion file consists of a symlink and then checks if the symlink can be resolved. In normal operation, this shouldn't cause troubles and all loops can be exited immediatly. When using a FreeBSD NFS share, these loops will make sure Lirc won't start before all configuartion files can be resolved. All with a timeout of 30 seconds in case of a network failure.